Add Kibana-alike custom shape library
This is a shape library for Diagrams app to create Mock-ups of
Kibana visualizations and dashboards. It contains most of the
visualizations and elements that are commonly used from Kibana
interface, using shapes from Mockup library and other Diagram
shape libraries.

This library includes an empty panel shape where other
visualization shapes can be arranged to create a mock-up for a
panel.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>